+++ CHANGELOG.md
@@ -7,6 +7,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3613 Memos on DBL settings in Devices view
 - YLP-3617 As a user, I can see the devices changes history
 - YLP-3785 Be able to know the effective DBLG1 or DBLG2 settings for the displayed day in the daily view
+- YLP-3813 Add hyperglycemia alerts (display in the patient view and settings in care teams and patients profile)
 
 ### Changed
 - YLP-2823 Use , instead of . for languages other than English
@@ -14,6 +15,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3634 Add tooltips for Rescue carbs and Manual bolus metrics
 - YLP-3815 Display update timestamp in yourloops
 - YLP-3830 Use a generic CGM icon for warmup events
+- YLP-3831 Alarms for G7 CGM
 - YLP-3834 Memo for Weight parameter
 
 ### Engineering